+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/cursors/FallbackCursor.java
@@ -25,7 +25,11 @@ import com.apple.foundationdb.record.RecordCoreException;
 import com.apple.foundationdb.record.RecordCursor;
 import com.apple.foundationdb.record.RecordCursorResult;
 import com.apple.foundationdb.record.RecordCursorVisitor;
+import com.apple.foundationdb.record.logging.KeyValueLogMessage;
+import com.apple.foundationdb.record.logging.LogMessageKeys;
 import com.apple.foundationdb.util.LoggableException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -54,6 +58,8 @@ import java.util.function.Function;
  */
 @API(API.Status.EXPERIMENTAL)
 public class FallbackCursor<T> implements RecordCursor<T> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(FallbackCursor.class);
+
     @Nonnull
     private final Function<RecordCursorResult<T>, RecordCursor<T>> fallbackCursorSupplier;
     @Nonnull
@@ -113,6 +119,9 @@ public class FallbackCursor<T> implements RecordCursor<T> {
                     inner.close();
                     inner = fallbackCursorSupplier.apply(lastSuccessfulResult);
                     nextResultFuture = inner.onNext();
+                    if (LOGGER.isInfoEnabled()) {
+                        LOGGER.info(KeyValueLogMessage.of("fallback triggered", LogMessageKeys.MESSAGE, throwable.getMessage()));
+                    }
                 }
                 alreadyFailed = true;
             }